minijail: Preserve the die() message in the stack
This change allocates a buffer in the stack where the die() message will
be formatted into. This causes the minidump to contain whatever crash
message was emitted to be captured in the file. The default behavior in
glibc's vsyslog(3) is to allocate a buffer somewhere in the heap, and
that's why we don't currently have these messages. bionic's vsyslog(3)
implementation does allocate a stack buffer, so we'll have N+1
redundancy of the log message.

Bug: 111271997
Test: Inspected the generated code, was convinced that the minidump
      should contain the crash message.
Change-Id: I4577c311d0d95bf44d90febfec09fbd06b9e35d1