Update patch set 3
Patch Set 3: Presubmit-Verified+1

=== Finished presubmit run: L25300000189669968 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/715442?revision=3&workplanId=L25300000189669968




TreeHugger finished with: 15 passed.





WARNING: The tip-of-tree(4886454) for git_master/cf_x86_phone-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4886206), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




WARNING: The tip-of-tree(4886445) for git_master/gce_x86_phone-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4886206), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.











WARNING: The tip-of-tree(4886445) for git_master/marlin-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4886214), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




WARNING: The tip-of-tree(4886454) for git_master/sailfish-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4886260), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



Congratulations!

Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 3
Tag: autogenerated:TreeHugger:Verify-Pass
Reviewer: Gerrit User 1062513 <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0
Label: Presubmit-Verified=+1