Update patch set 4
Patch Set 4:

> The problem with scmp_bpf_disasm is that it's not library-friendly :( It's only built as a standalone binary.

i don't see the problem.  this tool isn't a lib either.  scmp_bpf_disasm can read from stdin, so if we read the policy file, then created a pipe, forked+ran that tool, wrote the bpf to the pipe, then it would be transparent to the user.

i'm just not seeing that much value in improving this disasm code that in reality is used on the order of once a year when there is a more commonly used library out there.

Patch-set: 4
Reviewer: Gerrit User 1040868 <1040868@85c56323-6fa9-3386-8a01-6480fb634889>