minijail0: change default mount flags with -k
Lets change the default mount flags for -k to be more secure by default
rather than require everyone to remember to specify these all the time.
If people actually want dev+exec+suid, they can use an explicit 0 flag.

This also updates the libminijail C API so that flags=0 is changed to
secure defaults.  This means there's not an easy way for the callers
to actually get flags=0, but maybe that's a good thing.  If they need
that, they could hack around it with something like MS_SILENT, and we
can wait for them to file a feature request.

Bug: chromium:873216
Change-Id: Ia70b62cbb308eddfb58de3454fde3bfbac8e2f92