minijail: Handle cap_get_flag(3) failing with EINVAL
This change adds handling EINVAL on cap_get_flag(3). This is because
libcap rejects setting / getting any capabilities that it wasn't
compiled with, despite there being some wording on libcap(3)'s
cap_from_text(3) manpage that would lead the readers to think that all
the library would allow doing so.

Bug: None
Test: # ./minijail0 -T static --ambient \
        -c 'cap_dac_read_search,cap_dac_override+e' -- \
        /bin/true  # On Chrome OS.
Change-Id: Iea9b6737b208a201868731d44724eb1caa165fc5