minijail: Enter all namespaces before calling execve(2)
This change makes it so that all of the processes inside the namespace
(including the minijail-init process) are contained in the same set of
namespaces. This allows the use of nsenter(1) and setns(2) with the PID
that is written to a file using the -f flag to minijail0, or the
minijail_run_pid_*() family of functions.

Bug: None
Test: make tests
Change-Id: I3d546c7aeeaa3c875f41ce80b3504488020d2a5b