syscall_filter: Introduce the arch metadata attribute
This change introduces metadata attributes, which are a list of
key/value pairs that come after the policy entry.

It also adds the 'arch' metadata attribute. When present, if the
currently running architecture name is not present in the
comma-separated list of architectures in the 'arch' metadata attribute's
value, the line is skipped altogether. This allows filter policy authors
to have a single file that can work in multiple architectures.

Bug: None
Test: make tests

Change-Id: I4fe2b4f0ec52121ecdb6c7b312b6afaa62d3b36c