Update patch set 1
Patch Set 1:

Started make tests, will see if there are any problems related to this.

The specific minijail test was built and ran successfully on Android.

If libbase is not to be available for ChromeOS, this change will not go very far in simplifying maintenance. If so, not worth doing it IMHO.

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>