Update patch set 1
Patch Set 1: Code-Review+1

Is it possible to add a test for these two? Maybe it's going to be in one of those ones that are not supported in the automated systems due to the restrictions with the unprivileged namespaces :( but at least we would have caught this in less than 2 years.

Anyways, the change looks good as-is, giving it the highest vote I can due to not being part of the +2-giving ACLs anymore (maybe I should apply for an exception).

Patch-set: 1
Reviewer: Gerrit User 1066743 <1066743@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1