Allow skipping RO /proc mount.
Some process trees we'd like to sandbox (like Chrome) expect to be able
to write to /proc, e.g. to set /proc/<pid>/setgroups.

Allow users of the Minijail API to skip remounting of /proc RO, so that
process trees that expect to do their own sandboxing can still be
restricted with Minijail.

We probably don't need to expose this in the cmdline tool, since in
general these more complex sandbox invocations are using the library
directly.

Bug: chromium:928764
Test: Compiles.
Change-Id: Id52ef80b30e3467adcf88ddb4f3290556c0a5f87