Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > I don't know how I feel about this. If we allow a default action of "ACCEPT", we no longer have an allowlist-based policy. And our policies being allow-lists rather than block-lists have served us pretty well in the past.
> > 
> > So my instinct is to leave default action to being kill-process, or at least preventing some ret values from being accepted default actions.
> 
> yeah, that's fair. this i mostly used to do '@default log' during development, but explicitly disallowing '@default allow' seems fine.

@default log is also dangerous, since it's essentially LOG+ALLOW.

I would prefer for RET_LOG to be a cmdline argument and not a policy language option (the backend implementation can of course remain).

So, the options for @default should themselves be an allow-list of RET values that don't result in the policy itself becoming a block-list.

Patch-set: 4
Reviewer: Gerrit User 1076138 <1076138@85c56323-6fa9-3386-8a01-6480fb634889>