tools/compile_seccomp_policy: Optimize AST visitation
Since the AST nodes form a graph (and NOT a tree), the Block.accept()
method was running into a combinatorial explosion of calls to
Block.accept(). Some of the Visitor.visit() methods had some logic to
prevent processing a Block more than once, but it would still be invoked
multiple times.

This change adds a Visitor.visited() method so that Block.accept() can
stop unnecessary recursion, so that Block.accept() will be evaluated
exactly once per Visitor-Block pair, thus strengthening the previous
guarantee of Visitor.visit() being evaluated just once (and changing the
visitation guard to an assert). This also adds a unit test to prevent
this from regression.

Bug: chromium:856315
Test: ./tools/compiler_unittest.py

Change-Id: I79ba9f66ee63c51ada18175519c2cdab32efcd5b