tools/compile_seccomp_policy: Allow permissive default actions
This change provides an escape hatch for overriding the @default action
provided in the .policy file.

.policy files should always have restrictive default actions that
terminate the thread/process upon policy violation. But for debugging
purposes it is super convenient to generate a BPF file that just logs
any violations. So the --default-action flag was added to reconcile both
goals.

Bug: chromium:856315
Test: ./tools/compiler_unittest.py
Test: ./tools/parser_unittest.py
Test: ./tools/compile_seccomp_policy.py --default-action=log \
      test/seccomp.policy test/seccomp.bpf && \
      ~/libseccomp/tools/scmp_bpf_disasm < test/seccomp.bpf | grep LOG

Change-Id: Ib29964dfe21faeddeb6a92b827ffe33007e8dc55