Update patch set 3
Patch Set 3:

The main purpose of this is to give an example that folks can run on their desktops, mostly. That's why the comment makes it straightforward to run this directly out of the checkout.

It would be trivial to have an ARM policy though, since I can just test it on a Chromebook. So I can do that.

Patch-set: 3
Reviewer: Gerrit User 1076138 <1076138@85c56323-6fa9-3386-8a01-6480fb634889>