Update patch set 4
Patch Set 4:

I am still struggling a little bit with this.

On the one hand, policy files should prevent folks from shooting themselves in the foot. So removing permissive default actions is the right thing to do.

On the other hand, at least those could be checked with presubmit rules. If somebody tries to add a binary filter generated with --default-action=log, that will be significantly harder to catch.

Maybe we can add a mode that checks or lists what the default action of a binary filter is? And then we can use that in presubmit.

Patch-set: 4
Reviewer: Gerrit User 1076138 <1076138@85c56323-6fa9-3386-8a01-6480fb634889>