Update patch set 10
Patch Set 10:

We do care about hwasan, and it has been red on master branch for a week or two because of this.

ASan needs a dynamic (as opposed to compile-time) test because processes that start before /data is mounted run with non-instrumented libminijail.so.

HWASan may not require it. But it definitely needs at least a compile-time check!

Please keep the original author (i.e., me) in the loop when you partially revert a change.

Patch-set: 10
CC: Gerrit User 1005967 <1005967@85c56323-6fa9-3386-8a01-6480fb634889>