Update patch set 4
Patch Set 4:

(2 comments)

Done - the code will now always grab the mount flags of the source.
That was not the case before but I don't think it'll do any harm, WDYT?

Re build error, I don't know why that happens - sys/statvfs.h is included and works
on my machine, so I assume the constants are not defined on that machine?
I assume that's part of the libc, but the host seems to be using glibc as well.

Patch-set: 4
Reviewer: Gerrit User 1510663 <1510663@85c56323-6fa9-3386-8a01-6480fb634889>