Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> doesn't that mean minijail is going to break on real Android devices when it tries to read that file ?

In the latest PS we're not even trying to read the file.

But even in the previous patchset, the worst case scenario is that SELinux will block the read and the function will return false.

The reason for the latest PS is that TreeHugger does not allow landing CLs that generate previously-unseen SELinux denials. Given that Minijail is used by some processes that are started on boot, these processes are triggering a new denial for access to /proc/sys/kernel/seccomp/actions_avail. Nothing is breaking (the function is returning 0) but TreeHugger is blocking submit because of the extra denials.

I want to land this now and fix the denials in parallel, since it'll take me a while to get an Android environment working enough to test that the denials are gone. And as per the latest presubmit, PS 5 did actually fix things.

Patch-set: 5
Reviewer: Gerrit User 1076138 <1076138@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0