Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Hmmm actually the test failures look somewhat related:
> 
> NamespaceTest.test_bind_ro_with_pivot_root (run 1 of 2) (libminijail_unittest_gtest) 	00:00.000
>  No test results.
> libminijail[5322]: failed to fork child

At least that's the test that I added - The logs only seem to contain this line though, is there a way to properly debug this?
Without being able to reproduce this it's hard to fix it. I think a first step would be to
replace the die("failed to fork") with pdie()

Patch-set: 5
Reviewer: Gerrit User 1510663 <1510663@85c56323-6fa9-3386-8a01-6480fb634889>