Update patch set 1
Patch Set 1:

(setting scope)
the only time we respect the remount_mode is when mount namespaces are enabled.  this is enforced inside of libminijail itself (by only processing remount_mode when flags.vfs is true).  if we processed the remount_mode w/out requiring mount namespaces, that'd corrupt the current mount namespace, and we'd never want users to do that.

the reason for the CLI check is to make sure people aren't using -K w/out -v because doing so doesn't make sense ... mount namespaces aren't enabled, so we're going to ignore the -K, so users shouldn't have passed -K in the first place (didn't need it), or they expected that behavior, but it didn't happen, and now they're confused.

wrt "some options imply mount namespaces and the CLI check doesn't factor those in", i don't think this is a problem.  in the referenced https://crrev.com/c/1748736, and in other CrOS reviews like that, i argue that people *shouldn't* be relying on implicit -v when they actually do want -v.  i disagree with "-p implies -v therefore you should drop -v".  the implicit/internal behavior would be that if possible, and the daemons clearly want mount namespaces regardless of the -p setting, therefore they better be setting -v explicitly.  if it means a handful of devs passing -p & -K but not -v hit a failure, then so be it.

Patch-set: 1
Reviewer: Gerrit User 1040868 <1040868@85c56323-6fa9-3386-8a01-6480fb634889>