Update patch set 1
Patch Set 1:

Is this fixing an actual bug? The reason for the const_cast here is to be able to build the array properly, but bear in mind that minijail_run_pid_pipes takes a char* const [], which means the individual pointers will not be modified.

Patch-set: 1
Reviewer: Gerrit User 1076138 <1076138@85c56323-6fa9-3386-8a01-6480fb634889>