Update patch set 1
Patch Set 1:

> Is this fixing an actual bug? The reason for the const_cast here is to be able to build the array properly, but bear in mind that minijail_run_pid_pipes takes a char* const [], which means the individual pointers will not be modified.

I don't think there is or was a bug here although I'm not 100% sure. These const_casts and raw C-style casts feel like they're poking holes in the C/C++ type system. Something doesn't feel quite right.

Various Minijail functions take an array of strings as a char* const [], which is an array of const pointers to (implicitly mutable) C-style strings. I get the const pointers in the middle, but I'm wondering about the mutable C-style strings. The constness of pointers is not transitive and doesn't extend to the C-style strings they're pointing to. This partial but not deep constness is confusing.

I couldn't find any indication that these strings are actually getting modified and need to be mutable. Am I missing something? If there is no modification, then why not take these strings by const char* const [], with full and deep constness? This would avoid the confusion and the const_casts.

But if these strings are modified, then these tests are not correct. In particular, these tests share some global strings and don't expect them to be modified.

Patch-set: 1