minijail: support ranges
When a range like 2, 3, 4 is listed then check
for (in effect) >= 2 && <= 4, rather than individually
testing each number.  If the range starts from zero then
no base comparison is performed -- I'm starting to
2nd guess this particular optimization since
1) I'm not sure if the comparison is signed or
unsigned, and 2) I might've seen that some arches
actually do have syscalls in the 0x8000xxxx range.

Removing the optimization for >= 0 might also
reduce some of the complixities in the unit tests,
which I'm already debating how to make them work
across all arch's (unless there's some syscalls
with the same # on all).  An also is to allow
syscall #s to be explicitly specified in policy
files.

Bug: None
TEST=Ran and adjusted unit tests until they passed.

Change-Id: I46320d59444030bdcbe7ccc3eb77b83c4eea4858
Signed-off-by: Matt Delco <delco@chromium.org>