Update patch set 4
Patch Set 4:

We actually don't #define anything, but we test at compile time the architecture we're building for. Since the documentation for sys_clone explicitly specifies that the order of arguments depends on the architecture, I don't know how to avoid an #if defined() or an #ifdef somewhere, if we want to pass correct arguments in the right order to sys_clone.

Unfortunately, this fix doesn't prevent the SandboxedProcess unit tests from failing on ARM and ARM64 platforms. I'm a bit at a loss as to what could be causing this failure. Any idea?

Patch-set: 4