Update patch set 2
Patch Set 2: Verified+1

> Patch Set 1:
> 
> oh wait, i realized that the in* problem may affect other syscalls. created https://android-review.googlesource.com/c/platform/external/minijail/+/1168087 to fix that more generally.
> 
> you can make this change just about the kill syscall.

Ok, done.  I had to think for a second about why that other change doesn't cover kill, and it's because the other issues involve an eager match on a prefix of a word, while for "kill" it's matching on the whole word.

Patch-set: 2
Reviewer: Gerrit User 1588408 <1588408@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1