Add minijail_rlimit64
Add a variant of minijail_rlimit that takes rlim64_t as the type for the
`cur` and `max` parameters.  This is needed because the size of rlim_t
depends on whether you're using a 32-bit kernel or a 64-bit kernel.

This isn't a problem for most C and C++ programs because they get the
definition directly from the kernel header but it becomes an issue for
rust code.  The libc crate defines rlim_t as an alias for c_ulong so
when the code is compiled as a 32-bit executable the definition
conflicts with what all the other programs will use as the width of the
type.

Explicitly taking an rlim64_t avoid all this hassle because the width of
the type is independent of the architecture for which the program is
being compiled.

To make this work, change the existing implementation in libminijail to
use rlim64_t everywhere and make the old minijail_rlimit call forward to
minijail_rlimit64, after checking for RLIM_INFINITY.

Add `-D_LARGEFILE64_SOURCE` to the makefile so that the rlim64_t
definition will be available when compiling for a 32-bit userspace.

Bug: None
TEST=`tast run kevin vm.Virtiofs`

Change-Id: I98740c50b5e48881d1d69e6af5b4d413745c9d88