Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> wherever you define the FFI for libminijail, change the rlim_t to rlim64_t everywhere and you should be good to go
> 
> that said, it looks like it already does:
> io_jail/src/libminijail.rs:
>     pub fn minijail_rlimit(j: *mut minijail, kind: c_int, cur: u64, max: u64) -> c_int;
> 
> so what's the bug you're trying to fix ?

It does that because I spent a couple of hours figuring out why the rlimits were so wonky on kevin and then landed the change to replace the rlim_t in the FFI with u64.  If this is the Right Way (TM) to solve this problem then I guess I can just drop this change.  I just figured having a variant of this function that explicitly takes an rlim64_t would help other people avoid running into the same problem.


> 
> > Personally, I think that if we want to be able to call libminijail from languages other than C we need to be very explicit about the parameters that we are expecting.
> 
> welcome to the hell that is LFS.  the only sane solution is to build with LFS enabled everywhere which is what we do with common.mk.

:-/

Patch-set: 1
Reviewer: Gerrit User 1074798 <1074798@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0