generate_seccomp_policy: generate frequency files
The current generate_seccomp_policy.py can only generate policy files that
are sorted by frequency.  compile_seccomp_policy.py can accept a frequency file
(via a @frequency line in the policy file), but such a file isn't
created by generate_seccomp_policy.

This change adds a --frequency command-line option that can be used to
genreate a frequency file. For consistency I also added a --policy
option that can be used to specify the destination file for the policy
(normally the file is just sent to standard output and piped to a file).

If --frequency is used then the policy file will be generate syscall
rules in alphabetical order (vs. the frequency of the use).  Alphabetic
order makes it easier to see what syscalls are used and facilitates
manual modification.

I was tempted to also output an additional header comment to the
generated policy file (for both modes) that warn the files were
generated by a script (and to either add new syscalls in use order or to
the frequency file, as relevant for the mode that was used to create the
script).

TBD is whether to automatically add a @frequency line to the policy
file.  So far I haven't (the relative paths used to create the files
might have no relation to where the files are stored).  I thought about
adding a command-line parameter to compile_seccomp_policy.py that permits the
frequency file to be specified (instead of or in addition to @frequency) but
I didn't see a clean & easy way to add this.

Bug: None
Test: make tests
Change-Id: I9e1a5128052a7ae529b2e149c2e7bd4a263907a8