cli: implement --child-ld-preload(-keep)
--child-ld-preload <library.so>:
Allows setting LD_PRELOAD right before starting the target program,
without having it applied to minijail0's own invocation, which would be
the case if LD_PRELOAD was set manually in the environment before
calling minijail0.

The default behaviour of minijail is to clear LD_PRELOAD after the
program is started, when preload hooking is used (which is the default
for all dynamically linked programs). If LD_PRELOAD also needs to be
set for the program's descendants, we need --child-ld-preload-keep.

--child-ld-preload-keep:
If the target program is dynamically linked, preload hooking is used by
minijail0, injecting libminijailpreload.so in the target program.
After finishing the jail setup, the default behaviour is to clear
LD_PRELOAD from the environment before resuming the execution of
the program.
This implies that no children of the target program will
have LD_PRELOAD set.
This option tells libminijailpreload.so to keep LD_PRELOAD so that the
children of the target program inherit it.
This is never needed for statically linked programs, as preload hooking
is only possible with dynamically linked programs: in that case
LD_PRELOAD will be left untouched regardless of this option.

These two new switches (along with their counterpart in libminijail)
enable full control over LD_PRELOAD behaviour, on the minijail0 cli,
on the target program, and on the target program's descendants.

This commit removes the following todo from libminijailpreload.c:
   /* TODO(ellyjones): this trashes existing preloads, so one can't do:
    * LD_PRELOAD="/tmp/test.so libminijailpreload.so" prog; the
    * descendants of prog will have no LD_PRELOAD set at all.
    */