add parse_constants tool
NOTE: This demo version has implementations for gdb, gdb (MI), and
readelf.  The primary purpose is to allow people to compare the
3 different approaches (I've placed each in a separate file),
though they all have the inheriant problem of launching an app
and parsing the result.

---

This change adds a parse_constants tool that can be used to generate
a constants.json file from dump_constants by parsing the binary
contents of dump_constants (vs. the conventional method of running
dump_constants).  This alternative exists to assist with cross
compilation. This approach is akin generate_constants_json.py
though operations on ELF binaries rather than a compiler-specific
intermediate file. A tradeoff is that gdb or readelf must be installed.

An alternative would be to just add the constants.json file for
various architectures to the source tree.  Given the ABI guarantees
of the Linux kernel this would address most needs (except for new
syscalls or when new entries are added to gen_constants-inl.h).

In some environments 'gdb' will show the entire array contents by
just printing the array's name (e.g., "print syscall_table"). In
others this will only show the address of the array (placing
a * in front only shows the first entry, and using sizeof() says
the result has been optimized away).  To help deal with the latter
case I've modified the source to include the length of the array
in another variable (e.g., "syscall_table_len") and fallback to
using this length of necessary (e.g., "print syscall_table_len"
then "print *syscall_table@$1").

A tricky case is the arch information. This information can be parsed
out of variables, so I've added the arch information in a variable
(opting to use a "constant_entry" array to maximize reuse of the code
that's used to parse the syscall and constant tables).

An example use case would be for CrOS, where "parse_constants" would
be compiled for the host builder.  Each board would compile a
"dump_constants" and use "parse_constants dump_constants >
constants.json" to generate constants.json.

Bug: None
Test: make all.  Ran parser against dump_constants binaries that
are compiled for x64, arm32, and arm64.  Observed some potential
issues with sign extension of negative values, though the same behavior
is observed when running the x64 version of "dump_constants".

Change-Id: I753086692e2eb73b031eac7262f1c5441d2cd790