add parse_constants tool
This change adds a parse_constants tool that can be used to generate
a constants.json file from dump_constants by parsing the binary
contents of dump_constants (vs. the conventional method of running
dump_constants).  This alternative exists to assist with cross
compilation. This approach is akin generate_constants_json.py
though operations on ELF binaries rather than a compiler-specific
intermediate file.

A tricky case is the arch information. This information can be parsed
out of variables, so I've added the arch information in a variable
(opting to use a "constant_entry" array to maximize reuse of the code
that's used to parse the syscall and constant tables).

An example use case would be for CrOS, where "parse_constants" would
be compiled for the host builder.  Each board would compile a
"dump_constants" and use "parse_constants dump_constants >
constants.json" to generate constants.json.

Bug: None
Test: make all.  Ran parser against dump_constants binaries that
are compiled for x64, arm32, and arm64.  Observed some potential
issues with sign extension of negative values, though the same behavior
is observed when running the x64 version of "dump_constants".

Change-Id: Ibfd0c1d018a5f1faa7e062984dee2b1048156455