Update patch set 1
Patch Set 1:

I'm posting this to see how horrified of a reaction I get, though it is a legitimate attempt to be able to use pre-built bpfs in cross-compiled environments that don't necessarily use llvm.  The main other approaches I have are to limit pre-compiled use for platforms that match the host compiler (i.e., x64) or try to check in pre-compiled constants.json somewhere.

Patch-set: 1