Update patch set 1
Patch Set 1:

(4 comments)

> Patch Set 1:

> ha, I originally thought about implementing this through gdb and had the exact same reservation about what reaction it would have gotten, and that was what caused me to implement it with llvm ><.
> 
> one thing i _did not_ try was to use the output of gcc -S since it was platform-dependent and also could change, but each architecture seemed easy to parse by itself. another thing i did not try was to use GCC's RTL (gcc -fdump-lang-raw -S ...). it seemed a bit messier than llvm's IR, though, but I vaguely remembered that it was "simple" enough to parse quickly.
> 
> in any case, thanks for bringing support for pre-built bpf policies to CrOS!

Thanks for taking a look.  Offhand it was hard to know which of these representations are least likely to change over time.

Another approach I considered was using dumpelf, or similar, to get the symbol address/length and the segment's file offset to determine which bytes in the ELF file corresponding to the array.  It still would have the fragility of launching an app twice and parsing the output.  Otherwise it'd trade the fragility of parsing gdb's array output with the fun (endian and 32 vs. 64-bit type) of interpreting the raw bytes in the ELF.

I didn't spend much time looking at GDB's Python API.  The libgdb API doesn't seem to be very popular, and its API isn't much friendlier than passing raw commands to gdb.

Patch-set: 1