Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> this seems incredibly fragile/complicated and really seems like a fundamentally different approach is required

Yeah, I'd love it if posix moved away from doing things using scripts/apps and provided more programmatic APIs for performing tasks.  I'm pretty sure Windows has programmatic APIs for both debug symbols and parsing PE executables.

The most robust thing is probably to check in arch-specific constants.json files, given the fragilities in cross-compiling across multiple compiles and build environments.  I thought that a tool for parsing the ELFs would be the least compiler-specific approach.  The tradeoff is in who/what updates these json files (or whether to settle for some pre-commit script that tries to auto-generate the files and complains when the committed versions don't match--this is handy if a proposed commit has caused a change, though annoying if the delta is instead because of a change in the kernel or toolchain).

Patch-set: 1