Add more arm64 syscalls to constants.json
While trying to port crosvm to use the build-time compiler I ran
into complaints that certain basic syscalls like recv() and dup2()
aren't recognized in the arm64 build.

This change adds a few #define's so the syscalls are available.
I haven't tested whether the syscalls are used in crosvm, nor how
the runtime compiler is dealing with this (e.g., can it recognize
the syscalls because of how it's compiled, or it is complaining at
runtime and nobody's noticing).

On arm there's some syscalls that are only defined when compat
is requested, and others that are only defined when it's not.
So, I have to #include the header twice to get both.

Bug: None
Test: make all.  Ran dump_constants on arm64 with and without change
and noticed ~40 syscalls appear with this change.
Change-Id: I1aa99606186f9b56d58e2fb9c175ad5e5cc1fef2