cli: implement --child-ld-preload
--child-ld-preload <library.so>:
Allows setting LD_PRELOAD right before starting the target program,
without having it applied to minijail0's own invocation, which would be
the case if LD_PRELOAD was set manually in the environment before
calling minijail0.

The default behaviour of minijail is to clear LD_PRELOAD after the
program is started, when preload hooking is used (which is the default
for all dynamically linked programs).
When this option is used however, the preload library will not clear
LD_PRELOAD, hence all of the program's descendants will inherit it.

This commit removes the following todo from libminijailpreload.c:
   /* TODO(ellyjones): this trashes existing preloads, so one can't do:
    * LD_PRELOAD="/tmp/test.so libminijailpreload.so" prog; the
    * descendants of prog will have no LD_PRELOAD set at all.
    */

Change-Id: I2e106003824ff6707b80b6653141657e4ec40f47