Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> __ARCH_WANT symbols aren't meant for userspace code to define.  these syscalls aren't exported on arm64 because you're not supposed to be using them.  the C lib code (e.g. glibc) is supposed to implement dup2() on top of the dup3 syscall.  if crosvm is writing low level syscalls, it should be doing the samething.

It's Rust, and I think Rust's compiler is implemented in Rust, so offhand I don't know whether that means it 1) never uses glibc, 2) uses a glibc from when the current Rust compiler was made, or 3) uses a glibc from whichever Rust compiler was the last to not be implemented in Rust.

At the end I've listed the syscalls that I noticed crosvm using and none seem to overlap with the syscalls that are present in the policy file but unknown to a unmodified minijail lib.  I didn't poke into any of the Rust libraries used by crosvm.  Some basic symbol querying in gdb seems to show dup() and dup2() present, but not dup3() (granted, this doesn't catch cases via dlsym()).  I don't know if the arm64 policy files for crosvm were generated by hand (e.g., ported from x64), by using strace, or by running guess-and-check on the policy files until a jailed crosvm managed to run.

For minijail, I was mostly trying to enable the broadest set of syscalls without passing judgement on the quality/wisdom apps that might unwisely try to use certain syscalls.  One possible issue is what to do about some syscall names that can expand to different #s on arm64:

sendfile: 71 or 1046
ftruncate: 46 or 1047
truncate: 45 or 1048
stat: 1038 or 1049
lstat: 1039 or 1050
fstat: 80 or 1051
fcntl: 25 or 1052
fadvise64: 223 or 1053
newfstatat: 79 or 1054
fstatfs: 44 or 1055
statfs: 43 or 1056
lseek: 62 or 1057
mmap: 222 or 1058

Basically the first # is typical, but the second can be used if __SYSCALL_COMPAT and __ARCH_WANT_SYSCALL_OFF_T are defined (neither are defined in this proposed change).  The most unambiguous thing would be to cite the names with a "64" suffix, though this might confuse people as to why their code uses, e.g., "sendfile" but the policy file states "sendfile64".  I don't know how strace reports the names in its traces.

crosvm syscall() calls:

SYS_clone		220
SYS_getdents64		61
SYS_getegid[32]		177
SYS_geteuid[32]		175
SYS_getpid		172
SYS_memfd_create	279
SYS_renameat2		276
SYS_setresgid[32]	149
SYS_setresuid[32]	147

Patch-set: 2