Update patch set 2
Patch Set 2: Code-Review-1

> It's Rust

Rust is a language, not a C library.

> and I think Rust's compiler is implemented in Rust

the language that the compiler itself is written in is irrelevant to what it outputs.  we only care about its output & runtime.

> so offhand I don't know whether that means it 1) never uses glibc

both def use the system C library which is glibc.  it's fairly trivial to detect -- just run ldd(tree) on the binaries.  the programs directly link against the system C lib.

> 2) uses a glibc from when the current Rust compiler was made, or
> 3) uses a glibc from whichever Rust compiler was the last to not be implemented in Rust.

that isn't what Rust does, and there is no way we'd allow that kind of craziness in the first place for any language.

> At the end I've listed the syscalls that I noticed crosvm ... I don't know if the arm64 policy files for crosvm were generated by hand (e.g., ported from x64), by using strace, or by running guess-and-check on the policy files until a jailed crosvm managed to run.

sounds like you should do some research first on the quality of the filter and get back to us.

> Some basic symbol querying in gdb seems to show dup() and dup2() present, but not dup3()

those aren't syscalls, those are C library calls.  the C library is free to implement the call however it wants behind the scenes using whatever syscalls it likes.  and that's what glibc does for asm-generic ports (i.e. aarch64):

https://sourceware.org/git/?p=glibc.git;a=blob;f=sysdeps/unix/sysv/linux/generic/dup2.c;h=32592f9450415775ee816e2065eff43e7ed21cb2;hb=279c68ce1336d84d82ce491a4b77086e574ba380

> For minijail, I was mostly trying to enable the broadest set of syscalls without passing judgement on the quality/wisdom apps that might unwisely try to use certain syscalls.

minijail should not be exposing or relying on internal guts of headers when the arch porter clearly did not intend for it to do so, and when the code is clearly marked as "deprecated & likely to die soon".  in fact, it looks like they're doing just that already:

https://chromium.googlesource.com/chromiumos/third_party/kernel/+/a0673fdbcd42105261646cd4f3447455b5854a32%5E%21/

> One possible issue is what to do about some syscall names that can expand to different #s on arm64

i'm pretty sure you're getting confused by the compat layers that exist in the kernel and don't apply to aarch64 (because it's new enough to not include the legacy 32bit-64bit junk).  that is why all of these defines have __ prefixes -- they're meant for internal use only.

if you want to double check, build the arm64 kernel and then look at the arch/arm64/kernel/sys.[oci] files.  that will tell you with absolute certainly what syscalls are supported under arm64.

i'm of course ignoring arm32 & its syscall layer & runtime personality.  that has its own set of different warts, but has no real bearing on the arm64 syscall table.

> crosvm syscall() calls:

crosvm doesn't know about syscall numbers, just syscall names.  it leaves the build time headers to translate those names to the respective numbers.

Patch-set: 2
Reviewer: Gerrit User 1040868 <1040868@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1